fix data loading inconsistencies

+++ src/legend_lar/calibration/nre_c_inference.py
@@ -288,7 +288,7 @@ class NRECCalibrator:
         else:
             (_, f_idx, f_vals, ge_cu_seqlens, ge_max_seqlen, _) = hpge
             f_idx=f_idx.to(device=self.device, non_blocking=True).to(dtype=torch.long)
-            f_vals=f_vals.to(device=self.device, non_blocking=True).to(dtype=torch.long)
+            f_vals=f_vals.to(device=self.device, non_blocking=True).to(dtype=torch.float32)
             ge_cu_seqlens=ge_cu_seqlens.to(device=self.device, non_blocking=True).to(dtype=torch.long)
             ge_max_seqlen=int(ge_max_seqlen)
